[ios] Mod: Task #7542 float value like price, area update to string for keep the precision, area value limit length to 10

+++ source/ios/currant/Scene/Rent/Controller/CUTERentPriceViewController.m
@@ -83,7 +83,7 @@
 
 - (void)onDepositEdit:(id)sender {
     [self.form syncTicketWithBlock:^(CUTETicket *ticket) {
-        ticket.deposit = [CUTECurrency currencyWithValue:self.form.deposit.floatValue unit:self.form.currency];
+        ticket.deposit = [CUTECurrency currencyWithValue:self.form.deposit unit:self.form.currency];
     }];
 }
